requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.0,<3.13.0
dj-rest-auth[with_social]
drf-yasg[validation]
djangorestframework-simplejwt~=4.6.0